% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sheets_prefix.R
\name{add_sheets_prefix}
\alias{add_sheets_prefix}
\title{Add sheets name to REDCap exports}
\usage{
add_sheets_prefix(data, exept = NULL)
}
\arguments{
\item{data}{(data.frame) a "data" data frame inside the object
exported from REDCap using the \link{read_redcap} function.}

\item{exept}{(chr) Character vector of variable's names in \code{data} to
exclude from being considered into a sheet (i.e. REDCap
infromations not included into a specific form, e.g., ID)}
}
\value{
the data \link[tibble:tibble-package]{tibble} with modified
names.
}
\description{
Data exported from REDCap are extracted column-binding all the
sheets variable in order from left to right. When all the
variable of a sheet are considered, the column named
\verb{<sheet_name>_completed} is added to track the records for which
all the sheet is filled.
}
\details{
This function add the name of every sheets as a prefix of the name
of each of its corresponding variable (if not already present).
}
\examples{

sample_df <- data.frame(
  a = sample(1),
  b = sample(1),
  c_complete = sample(1),
  d = sample(1),
  e_foo = sample(1),
  e_complete = sample(1)
)


add_sheets_prefix(sample_df)
add_sheets_prefix(sample_df, exept = "a")
}
