% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_data_plot.R
\name{complete_data_plot}
\alias{complete_data_plot}
\title{Quality plot for completed data}
\usage{
complete_data_plot(data, summary_fun, type_completed)
}
\arguments{
\item{data}{data to use}

\item{summary_fun}{summary funciton (eg, \code{sum} or \code{mean})}

\item{type_completed}{(chr) what is represented (i.e. "Completed", or
"Non-Completed")}
}
\value{
ggplot2 plot
}
\description{
Quality plot for completed data
}
\examples{
purrr::iwalk(
  tipnet.ubesp:::generate_main_data(),
  ~ assign(.y, .x, pos = 1)
)

complete_data_plot(full_records, "sum", "Completed")
}
