% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-qualityReport.R
\name{completeDataPlot}
\alias{completeDataPlot}
\title{Quality plot for completed data}
\usage{
completeDataPlot(data, summary_fun, type_completed)
}
\arguments{
\item{data}{data to use}

\item{summary_fun}{summary funciton (eg, \code{sum} or \code{mean})}

\item{type_completed}{(chr) what is represented (i.e. "Completed", or
"Non-Completed")}
}
\value{
ggplot2 plot
}
\description{
Quality plot for completed data
}
