% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_from_server.R
\name{db_update_from_server}
\alias{db_update_from_server}
\title{Title}
\usage{
db_update_from_server(
  path_data = dir_create(here("data")),
  file_name = "tipnet.rds"
)
}
\arguments{
\item{path_data}{(chr) path for data folder (default is \verb{data/}
under the current project or the \link[here:here]{here} path, if it not
exists it will be created)}

\item{file_name}{(chr) name for the file to store the data}
}
\value{
(lgl) \code{TRUE} if success, \code{FALSE} otherwise
}
\description{
Title
}
\examples{
\dontrun{
  db_update_from_server()
}
}
