% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-missReport.R
\name{funs-missReport}
\alias{funs-missReport}
\alias{miss_dataToUse}
\alias{miss_dataPlot}
\alias{miss_dataTbl}
\title{functions for missing-data section}
\usage{
miss_dataToUse(.db)

miss_dataPlot(.db)

miss_dataTbl(.db, .center)
}
\arguments{
\item{.db}{(data frame) main data}

\item{.center}{(chr) center(s) to consider}
}
\description{
functions for missing-data section
}
\section{Functions}{
\itemize{
\item \code{miss_dataToUse}: data to use

\item \code{miss_dataPlot}: plot

\item \code{miss_dataTbl}: table
}}

