% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-outlierReport.R
\name{funs-outlierReport}
\alias{funs-outlierReport}
\alias{outlier_dataPlot}
\alias{outlier_dataToUse}
\title{Functions for outlier section}
\usage{
outlier_dataPlot(.db, .center)

outlier_dataToUse(.db, .center)
}
\arguments{
\item{.db}{(data frame) main data to use}

\item{.center}{(chr) name(s) of the center(s) to consider}
}
\description{
Functions for outlier section
}
\section{Functions}{
\itemize{
\item \code{outlier_dataPlot}: plot

\item \code{outlier_dataToUse}: data to use
}}

