% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-qualityReport.R
\name{funs-qualityReport}
\alias{funs-qualityReport}
\alias{quality_dataToUse}
\alias{quality_summaryFun}
\alias{quality_areOutAge}
\alias{quality_statGlobal}
\alias{quality_outOfAge}
\alias{quality_completeDataPlot}
\title{Quality module}
\usage{
quality_dataToUse(.db, completed)

quality_summaryFun(type)

quality_areOutAge(.db)

quality_statGlobal(.db, completed)

quality_outOfAge(are_out_age)

quality_completeDataPlot(.db, summary_fun, completed)
}
\arguments{
\item{.db}{(data frame) data}

\item{completed}{(chr) "Completed" or "Not-completed"}

\item{type}{(chr) "Total" or "Proportion"}

\item{are_out_age}{(lgl) logical vector of out of range values}

\item{summary_fun}{(chr) "sum" or "mean"}
}
\description{
Quality module
}
\section{Functions}{
\itemize{
\item \code{quality_dataToUse}: data to use

\item \code{quality_summaryFun}: summary function

\item \code{quality_areOutAge}: check for out of range

\item \code{quality_statGlobal}: message for global statistics

\item \code{quality_outOfAge}: message for out of age

\item \code{quality_completeDataPlot}: plot
}}

