% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_main_data.R
\name{generate_main_data}
\alias{generate_main_data}
\title{Global / extensive data setup}
\usage{
generate_main_data()
}
\description{
Here put data that should be cached at the server startup, and
computed only there once.
}
