% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sheet.R
\name{get_sheet}
\alias{get_sheet}
\title{Get a sheet table}
\usage{
get_sheet(x, sheet, field = NULL)
}
\arguments{
\item{x}{(tbl_df) output from \link{db_update_from_server}}

\item{sheet}{(chr) name of sheet to extract}

\item{field}{(chr, default = NULL) if not NULL, the name of the field
containing the sheet required. If NULL, it is supposed that there is only
one field containing the required sheet, and that is used. If more than
one field exist with the required sheet, and field is NULL, an error
is thrown suggesting the possible fields to consider.}
}
\value{
the required sheet \link[tibble:tibble-package]{tibble}
}
\description{
Get a sheet table
}
\examples{
\dontrun{
  tipnet <- read_rds(db_update_from_server())
  tipnet \%>\% get_sheet("anagrafica")
  tipnet \%>\% get_sheet("anagrafica", field = "anagrafica")

}
}
