% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_expired.R
\name{is_expired}
\alias{is_expired}
\title{Check if a file is expired}
\usage{
is_expired(filepath, max_days)
}
\arguments{
\item{filepath}{(chr) file path}

\item{max_days}{(int) number of days until expiration}
}
\value{
(lgl) \code{TRUE} if the file at \code{filepath} is expired respect
\code{max_days}. \code{FALSE} otherwise
}
\description{
Check if a file is expired
}
