% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-missReport.R
\name{module-missReport}
\alias{module-missReport}
\alias{missReportUI}
\alias{missReport}
\alias{missReportStatic}
\title{module for missing-data section}
\usage{
missReportUI(id, data)

missReport(id, data)

missReportStatic(data, center)
}
\arguments{
\item{id}{numeric vectors.}

\item{data}{data frame}

\item{center}{(chr) name(s) of the center(s) to consider}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{missReportUI}: user interface

\item \code{missReport}: server function

\item \code{missReportStatic}: static report function
}}

