% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-outlierReport.R
\name{module-outlierReport}
\alias{module-outlierReport}
\alias{outlierReportUI}
\alias{outlierReport}
\alias{outlierReportStatic}
\title{Outlier module}
\usage{
outlierReportUI(id, data)

outlierReport(id, data)

outlierReportStatic(data, center)
}
\arguments{
\item{id}{numeric vectors.}

\item{data}{data frame}

\item{center}{(chr) name(s) of the center(s) to consider}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{outlierReportUI}: user interface

\item \code{outlierReport}: server function

\item \code{outlierReportStatic}: static report function
}}

