% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-qualityReport.R
\name{module-qualityReport}
\alias{module-qualityReport}
\alias{qualityReportUI}
\alias{qualityReport}
\alias{qualityReportStatic}
\title{Quality module}
\usage{
qualityReportUI(id)

qualityReport(id, data)

qualityReportStatic(data, completed, type)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}

\item{completed}{(chr) "Completed" or "Not-completed"}

\item{type}{(chr) "Total" or "Proportion"}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{qualityReportUI}: user interface

\item \code{qualityReport}: server function

\item \code{qualityReportStatic}: static report function
}}

