% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tables.R
\name{nest_tables}
\alias{nest_tables}
\title{Extract field from REDCap db}
\usage{
nest_tables(data)
}
\arguments{
\item{data}{(list) the output of \link{tidy_extract} to extract the
data (not the meta-data).}
}
\value{
\link[tibble:tibble-package]{tibble}
}
\description{
Extract field from REDCap db
}
\examples{
\dontrun{
 tipnet_full <- read_redcap(url = tipnet_redcap_url())
 nest_tables(tipnet_full)
}
}
