% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-outlierReport.R
\name{outlierReport}
\alias{outlierReport}
\alias{outlierReportUI}
\title{Outlier module}
\usage{
outlierReportUI(id, data)

outlierReport(id, data)
}
\arguments{
\item{id}{numeric vectors.}

\item{data}{data frame}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{outlierReportUI}: user interface

User Interface description

\item \code{outlierReport}: server function

Server description
}}

