% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-qualityReport.R
\name{qualityReport}
\alias{qualityReport}
\alias{qualityReportUI}
\alias{qualityReportStatic}
\title{Quality module}
\usage{
qualityReportUI(id)

qualityReport(id, data)

qualityReportStatic(data, completed, type)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{qualityReportUI}: user interface

This define the User Interface side of the Shiny function. Note that
for every input used here involved in dynamically changes, there
should be a corresponding parameter on the header of the main
\code{index.Rmd} script, to permit the report to be rendered statically
too.

\item \code{qualityReport}: server function

This define the server side of the Shiny function. Note that for
every input used here involved in dynamically changes, there should
be a corresponding parameter on the header of the main \code{index.Rmd}
script, to permit the report to be rendered statically too.

\item \code{qualityReportStatic}: static report function

This should reproduce \code{qualityReport} for static production, ie
using i input the corresponding values last used by Shiny. They
should be pass to the \code{render} function as a parameter.
}}

