% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_readcap.R
\name{read_redcap}
\alias{read_redcap}
\title{Import the data from REDCap}
\usage{
read_redcap(url, token = tipnet_token())
}
\arguments{
\item{url}{(chr) the url for the REDCap database
(e.g.: "https://redcap.dctv.unipd.it/api/").}

\item{token}{(chr) Personal Access Token to the REDCap project}
}
\value{
(list) a list with 2 \link[tibble:tibble-package]{tibble}. The
first ("data") contains the data of the study, whereas the other
contains three columns with the names of the variables in \code{data},
the names of the form associated to each field and the labels
associated to each field
}
\description{
Import the data from REDCap repository into two
\link[tibble:tibble-package]{tibble}, one for the content and one
for the main metadata.
}
\note{
For \code{read_redcap()} to function properly, the user must have
Export permissions for the 'Full Data Set'.
}
\examples{
\dontrun{
  url   <- "https://bbmc.ouhsc.edu/redcap/api/"
  token <- "9A81268476645C4E5F03428B8AC3AA7B"
  rc <- read_redcap(token = token, url = url)
}

}
