% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-sampleReport.R
\name{sampleReport}
\alias{sampleReport}
\alias{sampleReportUI}
\title{Sample module}
\usage{
sampleReportUI(id)

sampleReport(id)
}
\arguments{
\item{id}{numeric vectors.}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{sampleReportUI}: user interface

User Interface description

\item \code{sampleReport}: server function

Server description
}}

