% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data_from.R
\name{tidy_data_from}
\alias{tidy_data_from}
\title{Tidy data from full REDcap extraction}
\usage{
tidy_data_from(x)
}
\arguments{
\item{x}{(lst) result of a call to \code{\link{read_redcap}}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Tidy data from full REDcap extraction
}
