% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_extract.R
\name{tidy_extract}
\alias{tidy_extract}
\title{Extract data from REDCap import in tidy way}
\usage{
tidy_extract(data, type = c("data", "meta"))
}
\arguments{
\item{data}{the output of \link{read_redcap}}

\item{type}{(chr) "data" to get the main data of the project, or
"meta" to get the meta-data.}
}
\value{
\link[tibble:tibble-package]{tibble}
}
\description{
After that information are retrieved by \link{read_redcap}, this
function extract data or metadata from it, cleaning variables
names, lowering every textual column, and renaming the event
field like "fields"
}
\examples{
\dontrun{
  tipnet_raw <- read_redcap(tipnet_redcap_url())

  tidy_extract(tipnet_raw, "data")
  tidy_extract(tipnet_raw, "meta")
}

}
