% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipnet.ubesp-package.R
\docType{package}
\name{tipnet.ubesp-package}
\alias{tipnet.ubesp}
\alias{tipnet.ubesp-package}
\title{tipnet.ubesp: TIPNet reports}
\description{
The goal of tipnet.ubesp is to implement and
    collect all the code needed for the development of the TIPNet
    reports by UBESP.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/tipnet.ubesp}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/tipnet.ubesp/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com} [copyright holder]

}
\keyword{internal}
