% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipnet_redcap_url.R
\name{tipnet_redcap_url}
\alias{tipnet_redcap_url}
\title{Get the REDCap's url for TIPNet}
\usage{
tipnet_redcap_url()
}
\value{
(chr) the REDCap url for the TIPNet data
}
\description{
Get the REDCap's url for TIPNet
}
\examples{
tipnet_redcap_url()
}
