% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_from_server.R
\name{db_update_from_server}
\alias{db_update_from_server}
\title{Pull REDCap data from the server}
\usage{
db_update_from_server(path_data = NULL)
}
\arguments{
\item{path_data}{(chr) path for data folder (default is \verb{data/}
under the current project or the \link[here:here]{here} path, if it not
exists it will be created)}
}
\value{
(lgl) \code{TRUE} if success, \code{FALSE} otherwise
}
\description{
Pull REDCap data from the server
}
\examples{
\dontrun{
  db_update_from_server()
}
}
