% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-qualityReport.R
\name{qualityReport}
\alias{qualityReport}
\alias{qualityReportUI}
\title{Quality module}
\usage{
qualityReportUI(id)

qualityReport(id, data)
}
\arguments{
\item{id}{numeric vectors.}

\item{data}{database to use}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{qualityReportUI}: user interface

User Interface description

\item \code{qualityReport}: server function

Server description
}}

